Switch to canvas shell on Rancher 2.9
